dure: give each supervisor concern its own file
The supervisor was one file holding startup and rollback, the state its
threads coordinate through, the relay, and a test suite covering all
three. Reading any one of them meant scrolling past the other two.
